modulariized the function to generate a color based on the user's name so that the user post and user signed in were the same.
