+++ frontend-runitup/src/utils/apiConfig.js
@@ -17,3 +17,12 @@ export const getAuthHeaders = () => {
     // Don't include 'Content-Type' here as it will be set automatically for FormData
   };
 };
+
+export const generateColor = (name) => {
+  let hash = 0;
+  for (let i = 0; i < name.length; i++) {
+    hash = name.charCodeAt(i) + ((hash << 5) - hash);
+  }
+  const c = (hash & 0x00ffffff).toString(16).toUpperCase();
+  return "#" + "00000".substring(0, 6 - c.length) + c;
+};
